+++ src/panels/AssetPanel.php
@@ -3,7 +3,6 @@
 namespace bedezign\yii2\audit\panels;
 
 use bedezign\yii2\audit\components\panels\DataStoragePanelTrait;
-use bedezign\yii2\audit\models\AuditData;
 use Yii;
 
 /**
@@ -33,20 +32,6 @@ class AssetPanel extends \yii\debug\panels\AssetPanel
             return parent::save();
         }
         return null;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function cleanup($maxAge = null)
-    {
-        $maxAge = $maxAge !== null ? $maxAge : $this->maxAge;
-        if ($maxAge === null)
-            return false;
-        return AuditData::deleteAll('type = :type AND created <= :created', [
-            ':type' => 'audit/asset',
-            ':created' => date('Y-m-d 23:59:59', strtotime("-$maxAge days")),
-        ]) !== false;
     }
 
 }