+++ src/panels/ConfigPanel.php
@@ -3,7 +3,6 @@
 namespace bedezign\yii2\audit\panels;
 
 use bedezign\yii2\audit\components\panels\DataStoragePanelTrait;
-use bedezign\yii2\audit\models\AuditData;
 use Yii;
 
 /**
@@ -22,20 +21,6 @@ class ConfigPanel extends \yii\debug\panels\ConfigPanel
         return Yii::$app->view->render('@yii/debug/views/default/panels/config/detail', [
             'panel' => $this,
         ]);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function cleanup($maxAge = null)
-    {
-        $maxAge = $maxAge !== null ? $maxAge : $this->maxAge;
-        if ($maxAge === null)
-            return false;
-        return AuditData::deleteAll('type = :type AND created <= :created', [
-            ':type' => 'audit/config',
-            ':created' => date('Y-m-d 23:59:59', strtotime("-$maxAge days")),
-        ]) !== false;
     }
 
 }