+++ src/panels/DbPanel.php
@@ -3,7 +3,6 @@
 namespace bedezign\yii2\audit\panels;
 
 use bedezign\yii2\audit\components\panels\DataStoragePanelTrait;
-use bedezign\yii2\audit\models\AuditData;
 use Yii;
 use yii\debug\models\search\Db;
 use yii\grid\GridViewAsset;
@@ -57,20 +56,6 @@ class DbPanel extends \yii\debug\panels\DbPanel
     public function registerAssets($view)
     {
         GridViewAsset::register($view);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function cleanup($maxAge = null)
-    {
-        $maxAge = $maxAge !== null ? $maxAge : $this->maxAge;
-        if ($maxAge === null)
-            return false;
-        return AuditData::deleteAll('type = :type AND created <= :created', [
-            ':type' => 'audit/db',
-            ':created' => date('Y-m-d 23:59:59', strtotime("-$maxAge days")),
-        ]) !== false;
     }
 
 }